Linked all elements inside modal_settings_container.js to functionality
main_ouput_container.js : Built system to set textareas with text obtained by the user on extension startup and stored inside localstorage. Initialized text styles to be set upon load based off of 'userSettings' in index.js.

modal_settings_container.js : Built the functionality to tie in the dropdown values being first set to current 'userSettings' and to register a changed value in the dropdown which will make changes to the textarea after the apply button is pressed.

index.css : Wrote code to add styling to the '_speed_reader_output_textarea' that was added so text could be displayed for the speed reader.

index.js : Added all of the currently changeable setting options the user will have access to inside the 'userSettings' map and added the ability to let the user open index.html and have it load properly by saving default text into localstorage.

manifest.json : Made an alteration to the "name" to fit the project.

popup.html : Made changes to existing code to give ids to the elements and be referenced from popup.js

popup.js : Wrote the functionality to take the input from the user from the textarea in popup.html and store the data in localstorage and have the 'ok_button' launch the actual application in another tab.